highway/pyproject.toml
[tool.poetry]
name = "highway"
version = "0.1.0"
description = ""
authors = ["Georgii Savin <sawin.georgij@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.0"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
sqlalchemy = "^2.0"
asyncpg = "^0.29"
alembic = "^1.13"
google-genai = "^1.18.0"
pillow = "^10.1.0"
motor = "^3.6.0"
langchain = "^0.3.17"
langchain-core = "^0.3.58"
langchain-community = "^0.3.16"
langchain-google-genai = "^2.1.4"
langgraph = "^0.4.8"
google-cloud-aiplatform = "^1.105.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"
debugpy = "1.8.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
